misc/HtmlTestSuite/results/test-1111.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>HTML Test: focus - key events</title>
<meta content="long" name="timeout"/>
<link href="http://www.intel.com/" rel="author" title="Intel"/>
<link href="https://html.spec.whatwg.org/multipage/#focus" rel="help"/>
<meta assert="flag" content="interact"/>
<meta assert="assert" content="Check if the key events received by document are targeted at the element when no element is focused"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/testdriver.js"/>
<script src="/resources/testdriver-vendor.js"/>
</head><body><div id="log"/>
<script>

var t1 = async_test("The keydown event must be targeted at the body element"),
    t2 = async_test("The keypress event must be targeted at the body element"),
    t3 = async_test("The keyup event must be targeted at the body element");

setup({timeout: 10000});

document.onkeydown = t1.step_func_done(function(evt){
  assert_equals(evt.target, document.body, "The keydown events must be targeted at the document's body.");
});

document.onkeypress = t2.step_func_done(function(evt){
  assert_equals(evt.target, document.body, "The keypress events must be targeted at the document's body.");
});

document.onkeyup = t3.step_func_done(function(evt){
  assert_equals(evt.target, document.body, "The keyup events must be targeted at the document's body.");
});

t1.step(function() {
  test_driver.send_keys(document.body, "a");
});

</script>
</body></html>
